Fix issue with passing token in API calls and split API calling file into adminApi and useApi
