+++ src/contains/public/Login.js
@@ -39,6 +39,7 @@ function Login() {
                 console.error('Error:', error);
             });
     };
+
     return (
         <form onSubmit={handleSubmit}>
             <label>
